requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.7.1
certifi==2025.8.3
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.3
colorama==0.4.6
constantly==23.10.4
coverage==7.13.1
cryptography==46.0.3
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.6.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-filter==23.2
django-health-check==3.20.8
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
ndg-httpsclient==0.5.1
packaging==25.0
paypalrestsdk==1.13.3
phonenumbers==9.0.21
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pusher==3.3.3
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.1
pyOpenSSL==25.3.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.1.1
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sentry-sdk==2.48.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==14.1.0
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
zope.interface==8.1.1
